refactor(utils): add utils.js in src
move matchAction and overNonNil in utils.js
